LPS-63096 Rename com.liferay.portal.portlet.container -> com.liferay.portal.osgi.web.portlet.container
