+++ marketplace/platform/bnd.bnd
@@ -92,7 +92,7 @@ Liferay-Releng-Dependent-Apps:\
 	com.liferay.portal.output.stream.container:1.0.0,\
 	com.liferay.portal.pop.notifications:1.0.0,\
 	com.liferay.portal.portlet.bridge.soy:1.0.0,\
-	com.liferay.portal.portlet.container.test:1.0.0,\
+	com.liferay.portal.osgi.web.portlet.container.test:1.0.0,\
 	com.liferay.portal.portlet.tracker:1.0.0,\
 	com.liferay.portal.rest.extender.test:1.0.0,\
 	com.liferay.portal.rest.extender:1.0.0,\
