+++ modules/portal-osgi-web/portal-osgi-web-portlet-container-test/bnd.bnd
@@ -1,5 +1,5 @@
 Bundle-Name: Liferay Portal OSGI WEB Portlet Container Test
-Bundle-SymbolicName: com.liferay.portal.portlet.container.test
+Bundle-SymbolicName: com.liferay.portal.osgi.web.portlet.container.test
 Bundle-Version: 1.0.0
 Include-Resource:\
 	test-classes/integration,\
