+++ modules/portal-osgi-web/portal-osgi-web-portlet-container-test/src/testIntegration/resources/com/liferay/portal/osgi/web/portlet/container/embedded/test/template/dependencies/portlet_display_template_layout.ftl
